Move stuff around. Making json support purely client side as it needs to work well with the `monaco` module :rose:
