+++ src/app/monaco/languages/json/README.md
@@ -2,10 +2,22 @@
 
 This folder contains a bunch of stuff needed to parse json mostly pulled off the internet. Origin sources are mentioned in the headers.
 
-Core Stuff:
+# Core Folder
+mostly copied with simple modifications:
 * core/jsonc-parser: The `scanner` really.
+* core/jsonParser: The fancy stuff. The `parser` + Validator (using a provided schema)
+* core/jsonLocation: just helps manage json paths (`foo.bar.bas` etc). Also helps match e.g. (`foo.*` etc)
+* core/localize: don't want to take a dependency on VSCode localize. So a simple replacement function
+* core/jsonSchema: just the interface declaring the structure of json schemas.
 
-Monaco Stuff:
+# Services Folder
+Still stuff from VSCode but heavily modified to use the above ^ to give intelligence based on how I wanted it
+
+* `jsonSchemaService.ts`
+* `jsonCompletion.ts` file is the one I needed to customize to get what I wanted for our types.
+
+
+# Monaco Stuff:
 * monacoJson: The entry point for monacoJson stuff.
 * tokenization: Provides the syntax highlighting for monaco.
