+++ src/app/monaco/languages/json/core/jsonSchema.ts
