+++ src/app/monaco/languages/json/core/localize.ts
