+++ src/app/monaco/languages/json/service/schemas/package.json
