+++ src/app/monaco/languages/json/service/schemas/tsconfig.json
