+++ src/app/monaco/languages/json/utils/strings.ts
