+++ src/app/monaco/languages/json/utils/uri.ts
