--- src/server/lang/config/json/service/vscode-types.ts
@@ -1,89 +1,0 @@
-/**
- *
- *
- * From https://github.com/Microsoft/vscode-languageserver-node/blob/9c897747cbd0cbf58a5d1894377d1ed129698dcd/server/src/protocol.ts
- *
- *
- */
-
-/**
- * The kind of a completion entry.
- */
-export enum CompletionItemKind {
-	Text = 1,
-	Method = 2,
-	Function = 3,
-	Constructor = 4,
-	Field = 5,
-	Variable = 6,
-	Class = 7,
-	Interface = 8,
-	Module = 9,
-	Property = 10,
-	Unit = 11,
-	Value = 12,
-	Enum = 13,
-	Keyword = 14,
-	Snippet = 15,
-	Color = 16,
-	File = 17,
-	Reference = 18
-}
-
-/**
- * A completion item represents a text snippet that is
- * proposed to complete text that is being typed.
- */
-export interface CompletionItem {
-	/**
-	 * The label of this completion item. By default
-	 * also the text that is inserted when selecting
-	 * this completion.
-	 */
-	label: string;
-
-	/**
-	 * The kind of this completion item. Based of the kind
-	 * an icon is chosen by the editor.
-	 */
-	kind?: number;
-
-	/**
-	 * A human-readable string with additional information
-	 * about this item, like type or symbol information.
-	 */
-	detail?: string;
-
-	/**
-	 * A human-readable string that represents a doc-comment.
-	 */
-	documentation?: string;
-
-	/**
-	 * A string that shoud be used when comparing this item
-	 * with other items. When `falsy` the [label](#CompletionItem.label)
-	 * is used.
-	 */
-	sortText?: string;
-
-	/**
-	 * A string that should be used when filtering a set of
-	 * completion items. When `falsy` the [label](#CompletionItem.label)
-	 * is used.
-	 */
-	filterText?: string;
-
-	/**
-	 * A string that should be inserted a document when selecting
-	 * this completion. When `falsy` the [label](#CompletionItem.label)
-	 * is used.
-	 */
-	insertText?: string;
-
-	/**
-	 * An data entry field that is preserved on a completion item between
-	 * a [CompletionRequest](#CompletionRequest) and a [CompletionResolveRequest]
-	 * (#CompletionResolveRequest)
-	 */
-	data?: any
-}
